refactor: migrate to Nextjs Route Handler
