+++ packages/server/next/app/api/chat/config/llamacloud/route.ts
@@ -1,0 +1,32 @@
+import { getEnv } from "@llamaindex/env";
+import { LLamaCloudFileService } from "llamaindex";
+import { NextRequest, NextResponse } from "next/server";
+
+export async function GET(request: NextRequest) {
+  if (!getEnv("LLAMA_CLOUD_API_KEY")) {
+    return NextResponse.json(
+      {
+        error: "env variable LLAMA_CLOUD_API_KEY is required to use LlamaCloud",
+      },
+      { status: 500 },
+    );
+  }
+
+  try {
+    const config = {
+      projects: await LLamaCloudFileService.getAllProjectsWithPipelines(),
+      pipeline: {
+        pipeline: getEnv("LLAMA_CLOUD_INDEX_NAME"),
+        project: getEnv("LLAMA_CLOUD_PROJECT_NAME"),
+      },
+    };
+    return NextResponse.json(config, { status: 200 });
+  } catch (error) {
+    return NextResponse.json(
+      {
+        error: "Failed to fetch LlamaCloud configuration",
+      },
+      { status: 500 },
+    );
+  }
+}
